+++ src/index.ts
@@ -9,12 +9,14 @@ import minimist from 'minimist';
 import * as utils from './utils';
 import * as cookies from './cookies';
 import * as srr from './srr';
+import * as backfill from './backfill';
 
 
 const args = minimist(process.argv.slice(2), {
 
     // All options / switches
     'boolean': [
+        'backfill',
         'debug',
         'help',
         'login',
@@ -23,6 +25,7 @@ const args = minimist(process.argv.slice(2), {
 
     // Aliases for options
     alias: {
+        b: 'backfill',
         d: 'debug',
         h: 'help',
         l: 'login',
@@ -57,18 +60,42 @@ const args = minimist(process.argv.slice(2), {
         }
     }
 
+    else if (args.backfill) {
+        const validCookie = await cookies.checkLoginCookie()
+        if (!validCookie) {
+            console.log('Please login to srrdb.com...');
+            cookies.getLoginCookie();
+        }
+
+        backfill.processBackfill();
+    }
+
     // When no switch consider every parameter as file to upload
     // https://github.com/substack/minimist#var-argv--parseargsargs-opts
     else if (args._) {
-        // User needs to log in first.
+        // User needs to log in first. We don't check if the cookie is still valid here to save unnecessary calls to srrdb.
         const cookie = process.env.COOKIE || '';
         if (cookie === '') {
             console.log('No existing login cookie found, please login to srrdb.com first...');
             await cookies.getLoginCookie();
         }
-        args._.forEach((file: string) => {
-            srr.srrUpload(file);
-        });
+
+        let lastUploadSuccessful = false;
+
+        // Call srrUpload for each file
+        for (const file of args._) {
+            const uploaded = await srr.srrUpload(file)
+            if (uploaded) {
+                lastUploadSuccessful = true;
+            } else {
+                lastUploadSuccessful = false;
+            }
+        }
+
+        // if the most recent file upload succeeded, take a look in backfill folder and process them
+        if (lastUploadSuccessful){
+            backfill.processBackfill();
+        }
     }