+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 
-## [v1.0.0](https://github.com/peps1/airdcpp-user-commands/tree/v1.0.0) (2020-08-01)
-[Full git log](https://github.com/peps1/airdcpp-user-commands/compare/f1295271029803754d14f0e0dc01eb5466a6863a...v1.0.0)
+## [v2.0.0](https://github.com/peps1/srrup/tree/v2.0.0) (2020-08-04)
+[Full git log](https://github.com/peps1/srrup/compare/v1.0.0...v2.0.0)
+
+### Added
+* backfill feature
+* new command line switch (-b / --backfill) to trigger backfill directly
+
+### Changed
+* Better check to see if srr upload failed
+
+
+## [v1.0.0](https://github.com/peps1/srrup/tree/v1.0.0) (2020-08-01)
+[Full git log](https://github.com/peps1/srrup/compare/40204681df9d18e5a376407564dc3585d2bcef3d...v1.0.0)
 
 Initial release